Modifed regression tests of EPANET and SWIMM.  created /test/regression_base.py for the base class of both SWMM and EPANET classes to reduce code redundency. Removed test_project under test/swmm and test/epanet, created swmmregressiontest.py and epanetregresstest.py under these directories. Both test classes inherit from the RegressionBase.  Moved some test methods to /test/section_match.py.  Regression test conducted for EPANET (all pass), SWMM (most do not pass, appears to be problem in extra sections written to the new inp files).
