+++ test/core/section_match.py
@@ -169,3 +169,142 @@ def match_two_lists(t_fields, a_fields, ignore_trailing_0):
         return False
     return True
 
+
+def get_immediate_subdirectories(a_dir):
+    import os
+    return [name for name in os.listdir(a_dir)
+            if os.path.isdir(os.path.join(a_dir, name))]
+
+
+def readline_nowhite(f):
+    """ Readline_nowhite
+     Strip all carriage returns, tabs and spaces
+     If it is empty line, read the next line
+     Return the lines read, and the line without white
+    """
+    f_line = f.readline()
+    line_no = 1
+    f_line_nw = f_line.replace('\n', '').replace('\t', '').replace(' ', '')
+
+    while f_line_nw == '' and f_line != '':
+        f_line = f.readline()
+        line_no += 1
+        f_line_nw = f_line.replace('\n', '').replace('\t', '').replace(' ', '')
+    f_line = f_line_nw
+    return line_no, f_line
+
+
+def compare_two_files(fname1, fname2, exempted_strings):
+    """Reference: http://www.opentechguides.com/how-to/article/python/58/python-file-comparison.html
+    Modified by xw 2016/05/12
+    Strip all white spaces, empty lines and exempt the lines with exempted_strings
+    Return a string message diff_msg. If identical, return empty string
+    """
+
+    diff_msg = ''
+    # Open file for reading in text mode (default mode)
+    f1 = open(fname1)
+    f2 = open(fname2)
+
+    # Initialize counter for line number
+    line_no_f1 = 0
+    line_no_f2 = 0
+
+    # Readline_nowhite
+    # Strip all carriage returns, tabs and spaces
+    # If it is empty line, read the next line
+    # File 1:
+    line_read, f1_line = readline_nowhite(f1)
+    line_no_f1 = line_no_f1 + line_read
+
+    # File 2:
+    line_read, f2_line = readline_nowhite(f2)
+    line_no_f2 = line_no_f2 + line_read
+
+    # Loop if either file1 or file2 has not reached EOF
+    while f1_line != '' or f2_line != '':
+
+        # Compare the lines from both file
+        if f1_line != f2_line:
+
+            is_diff = True
+            # Loop through lines with keyword(s) indicating an exemption
+            for exempted_string in exempted_strings:
+                str_nw = exempted_string.replace(' ', '')
+                if f1_line.find(str_nw) != -1 \
+                        and f2_line.find(str_nw) != -1:
+                    is_diff = False
+                    # Break when the exempted_string hit
+                    break
+
+            if is_diff:
+                diff_msg = '<br>[old]Line-' + str(line_no_f1) + ':' + f1_line + '<br>' + \
+                           '[new]Line-' + str(line_no_f2) + ':' + f2_line + '<br>'
+                # Break on the first difference
+                break
+
+        # File 1:
+        line_read, f1_line = readline_nowhite(f1)
+        line_no_f1 = line_no_f1 + line_read
+
+        # File 2:
+        line_read, f2_line = readline_nowhite(f2)
+        line_no_f2 = line_no_f2 + line_read
+
+    # Close the files
+    f1.close()
+    f2.close()
+    return diff_msg
+
+
+def compare_two_analysis_blocks(fname1, fname2, exempted_strings,
+                                str_start="Analysis begun", str_end="Analysis ended"):
+    """Modified from method compare_two_files by xw 3/27/2017
+    Compare text between "Analysis begun" and "Analysis ended" of EPANET rpt files:
+    Strip all white spaces, empty lines and exempt the lines with exempted_strings
+    Return a string message diff_msg. If identical, return empty string
+    """
+
+    diff_msg = ''
+    # Open file for reading in text mode (default mode)
+
+    fnames = [fname1, fname2]
+    new_lines = [[], []]
+    try:
+        for ifile, fname in enumerate(fnames):
+            f = open(fname)
+            lines = f.readlines()
+            start_compare = False
+
+            for id_line, line in enumerate(lines):
+                if line.find(str_start) != -1:
+                    start_compare = True
+                elif line.find(str_end) != -1:
+                    start_compare = False
+                elif start_compare:
+                    if line.replace(' ', '').replace('\t', '').replace('\n', '') != '':
+                        is_exempted = False
+                        for exempted_string in exempted_strings:
+                            if line.find(exempted_string) == -1:
+                                is_exempted = False
+                            else:
+                                is_exempted = True
+                                break
+                        if is_exempted is False:
+                            new_lines[ifile].append([id_line, line])
+                else:
+                    pass
+            f.close()
+    except IOError as e:
+        diff_msg = '<br>IO Error: {} <br>'.format(str(e))
+
+    for f1_line, f2_line in zip(new_lines[0], new_lines[1]):
+        if f1_line[1] != f2_line[1]:
+            is_diff = True
+            if is_diff:
+                diff_msg = '<br>[old]Line-' + str(f1_line[0]) + ':' + f1_line[1] + '<br>' + \
+                           '[new]Line-' + str(f1_line[0]) + ':' + f2_line[1] + '<br>'
+                # Break on the first difference
+                break
+    return diff_msg
+
