+++ test/core/unittest_all_with_coverageapi.py
@@ -21,15 +21,17 @@ if __name__ == "__main__":
 
     # Create a coverage instance
     cov = coverage.Coverage()
+    cov.exclude('^\s*(import|from)\s')  # exclude import statements
+    # cov.exclude('^\s*(class)\s')  # regex does not work for class names
+
     cov.start()
 
     runner.run(my_suite)
+    fp.close()
 
     cov.stop()
     cov.save()
     cov.html_report()
-
-    fp.close()
 
     # Open coverage report
     current_path = os.getcwd()
